pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yoto-library"
version = "0.3.0"
description = "Manage Yoto CYO playlists as folders on disk"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Steve Rice" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio",
]
dependencies = [
    "click>=8.0",
    "httpx>=0.27",
    "keyring>=25.0",
    "numpy>=1.0",
    "Pillow>=10.0",
    "openai>=1.0",
    "together>=1.0",
    "google-genai>=1.0",
    "anthropic>=0.40",
    "python-dotenv>=1.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.0",
    "pytest-tmp-files>=0.0.2",
    "ruff>=0.11",
    "ty>=0.0.1a7",
    "iterm2>=2.0",
    "commitizen>=4.0",
    "cz-conventional-gitmoji>=0.7",
    "pre-commit>=4.0",
]
iterm2 = [
    "iterm2>=2.0",
]

[project.urls]
Homepage = "https://github.com/steverice/yoto-library"
Source = "https://github.com/steverice/yoto-library"
Issues = "https://github.com/steverice/yoto-library/issues"

[project.scripts]
yoto = "yoto_cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/yoto_lib", "src/yoto_cli"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "integration: tests requiring network access and external tools (yt-dlp)",
]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = [
    "E", "W", "F",          # pycodestyle + pyflakes
    "I", "TID", "TCH",      # imports: isort, tidy-imports, type-checking
    "UP", "FA", "YTT",      # modernization: pyupgrade, future-annotations, 2020
    "ANN",                   # type annotations
    "B", "SIM", "C4",       # bugs: bugbear, simplify, comprehensions
    "RSE", "PIE", "PERF",   # bugs: raise, pie, performance
    "RET",                   # return style
    "S",                     # security: bandit
    "BLE",                   # blind except
    "RUF", "PT",             # ruff-specific, pytest style
    "ERA",                   # eradicate (no commented-out code)
    "FLY", "ISC",            # f-strings, implicit string concat
    "T10", "T20",            # no debugger, no print
    "PTH",                   # pathlib over os.path
    "A", "SLF", "N",        # no builtin shadowing, no private access, PEP 8 naming
    "ASYNC",                 # async anti-patterns
    "PGH",                   # no blanket noqa/type:ignore
    "INP",                   # no implicit namespace packages
]

ignore = [
    "ANN204",   # __init__ return type is always None — obvious, no value
    "ANN401",   # Any is sometimes the right type (dict values, kwargs, JSON)
    # Project-specific:
    "S101",     # assert is valid for type narrowing
    "S603",     # subprocess calls are intentional (ffmpeg, mkvtoolnix, yt-dlp, osascript, claude CLI)
    "S607",     # subprocess calls use known tool names
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S105", "S106", "S108",  # fake secrets/paths in test fixtures
    "ANN",                    # test functions don't need type annotations
    "SLF001",                 # tests legitimately access private members
    "INP001",                 # test directories use pytest discovery, not __init__.py
    "TC",                     # tests use runtime imports freely
    "PTH",                    # tests can use open() directly
    "BLE001",                 # tests may catch broadly for skip/xfail
    "RET",                    # return style not important in tests
    "PERF",                   # performance rules not important in tests
]
"validation/**/*.py" = [
    "INP001",                 # standalone validation scripts, not a package
    "T201",                   # validation scripts use print() for output
]

[tool.ruff.lint.isort]
known-first-party = ["yoto_lib", "yoto_cli"]

[tool.ty.rules]
# Suppress unused-ignore-comment: ty resolves different imports depending
# on environment (local vs CI), so ignore comments needed in one context
# appear unused in the other.
unused-ignore-comment = "ignore"

[tool.ty.environment]
python-version = "3.10"
root = ["src"]

[tool.commitizen]
name = "cz_gitmoji"
version_provider = "pep621"
tag_format = "v$version"
update_changelog_on_bump = true

